common base
