+++ runtime/gc_stats/MarkJavaStats.hpp
@@ -24,37 +24,22 @@
 #if !defined(MARKJAVASTATS_HPP_)
 #define MARKJAVASTATS_HPP_
 
-#include "j9port.h"
-#include "modronopt.h"
-
-#include "Base.hpp"
 #include "AtomicOperations.hpp"
-#include "ReferenceStats.hpp"
+#include "JavaStats.hpp"
 
 /**
  * Storage for statistics relevant to the mark phase of a global collection.
  * @ingroup GC_Stats
  */
-class MM_MarkJavaStats : public MM_Base {
+class MM_MarkJavaStats : public MM_JavaStats {
 	/* data members */
 private:
 protected:
 public:
-	UDATA _unfinalizedCandidates; /**< unfinalized objects that are candidates to be finalized visited this cycle */
-	UDATA _unfinalizedEnqueued; /**< unfinalized objects that are enqueued during this cycle (MUST be less than or equal _unfinalizedCandidates) */
-
-	UDATA _ownableSynchronizerCandidates; /**< number of ownable synchronizer objects visited this cycle */
 	UDATA _ownableSynchronizerCleared; /**< number of ownable synchronizer objects cleared this cycle */
-
-	MM_ReferenceStats _weakReferenceStats; /**< Weak reference stats for the cycle */
-	MM_ReferenceStats _softReferenceStats; /**< Soft reference stats for the cycle */
-	MM_ReferenceStats _phantomReferenceStats; /**< Phantom reference stats for the cycle */
 
 	UDATA _stringConstantsCleared; /**< The number of string constants that have been cleared during marking */
 	UDATA _stringConstantsCandidates; /**< The number of string constants that have been visited in string table during marking */
-
-	UDATA _monitorReferenceCleared; /**< The number of monitor references that have been cleared during marking */
-	UDATA _monitorReferenceCandidates; /**< The number of monitor references that have been visited in monitor table during marking */
 
 #if defined(J9MODRON_TGC_PARALLEL_STATISTICS)
 	UDATA splitArraysProcessed; /**< The number of array chunks (not counting parts smaller than the split size) processed by this thread */
@@ -69,18 +54,10 @@ public:
 	void merge(MM_MarkJavaStats* statsToMerge);
 
 	MM_MarkJavaStats() :
-		MM_Base()
-		, _unfinalizedCandidates(0)
-		, _unfinalizedEnqueued(0)
-		, _ownableSynchronizerCandidates(0)
+		MM_JavaStats()
 		, _ownableSynchronizerCleared(0)
-		, _weakReferenceStats()
-		, _softReferenceStats()
-		, _phantomReferenceStats()
 		, _stringConstantsCleared(0)
 		, _stringConstantsCandidates(0)
-		, _monitorReferenceCleared(0)
-		, _monitorReferenceCandidates(0)
 	{
 		clear();
 	}
